refactor: make getObject utility reusabble
