+++ src/preview/preview-store.ts
@@ -140,99 +140,6 @@ export class PreviewStore<V> {
 		return this.metaDown;
 	}
 
-	public getModel(modelName?: Types.ModelName): Types.ModelSurface | undefined {
-		switch (modelName) {
-			case Types.ModelName.AlvaApp:
-				return Model.AlvaApp;
-			case Types.ModelName.Element:
-				return Model.Element;
-			case Types.ModelName.ElementAction:
-				return Model.ElementAction;
-			case Types.ModelName.ElementContent:
-				return Model.ElementContent;
-			case Types.ModelName.ElementProperty:
-				return Model.ElementProperty;
-			case Types.ModelName.Page:
-				return Model.Page;
-			case Types.ModelName.Pattern:
-				return Model.Pattern;
-			case Types.ModelName.PatternLibrary:
-				return Model.PatternLibrary;
-			case Types.ModelName.PatternProperty:
-				return Model.PatternProperty;
-			case Types.ModelName.PatternSlot:
-				return Model.PatternSlot;
-			case Types.ModelName.UserStore:
-				return Model.UserStore;
-			case Types.ModelName.UserStoreAction:
-				return Model.UserStoreAction;
-			case Types.ModelName.UserStoreEnhancer:
-				return Model.UserStoreEnhancer;
-			case Types.ModelName.UserStoreProperty:
-				return Model.UserStoreProperty;
-			case Types.ModelName.UserStoreReference:
-				return Model.UserStoreReference;
-			default:
-				return;
-		}
-	}
-
-	public getObject(constructorName: string, id: string): Model.AnyModel | undefined {
-		if (constructorName === 'Element') {
-			return this.project.getElementById(id);
-		}
-
-		if (constructorName === 'ElementAction') {
-			return this.project.getElementActionById(id);
-		}
-
-		if (constructorName === 'ElementContent') {
-			return this.project.getElementContentById(id);
-		}
-
-		if (constructorName === 'ElementProperty') {
-			return this.project.getElementPropertyById(id);
-		}
-
-		if (constructorName === 'Page') {
-			return this.project.getPageById(id);
-		}
-
-		if (constructorName === 'Pattern') {
-			return this.project.getPatternById(id);
-		}
-
-		if (constructorName === 'PatternLibrary') {
-			return this.project.getPatternLibraryById(id);
-		}
-
-		if (constructorName === 'Project') {
-			return this.project;
-		}
-
-		if (constructorName === 'UserStore') {
-			return this.project.getUserStore();
-		}
-
-		if (constructorName === 'UserStoreAction') {
-			return this.project.getUserStore().getActionById(id);
-		}
-
-		if (constructorName === 'UserStoreProperty') {
-			return this.project.getUserStore().getPropertyById(id);
-		}
-
-		if (constructorName === 'UserStoreEnhancer') {
-			return this.project.getUserStore().getEnhancer();
-		}
-
-		if (constructorName === 'UserStoreReference') {
-			return this.project.getUserStore().getReferenceById(id);
-		}
-
-		return;
-	}
-
 	public getProperties<T>(
 		element: Model.Element
 	): { [propName: string]: Types.ElementPropertyValue } {
